move processing of tex to views
