+++ bokehjs/src/lib/external/mathjax.d.ts
@@ -21,4 +21,5 @@ declare namespace MathJax {
   function tex2svg(formula: string, options?: ConvertOptions, macros?: TeXMacros): HTMLElement
   function ascii2svg(_formula: string): HTMLElement
   function mathml2svg(formula: string): HTMLElement
+  function find_math(text: string): ProtoItem[]
 }
