+++ bokehjs/src/lib/models/text/utils.ts
@@ -1,26 +1,6 @@
 import {TeX} from "./math_text"
 import {isString} from "core/util/types"
 import {BaseText} from "./base_text"
-import {find_math, tex2svg} from "./mathjax"
-
-export function process_tex_parts(text: string, math_parts: MathJax.ProtoItem[]): string {
-  const parts: string[] = []
-
-  math_parts.reduce((last_index = 0, math_part) => {
-    parts.push(text.slice(last_index, math_part.start.n))
-    parts.push(tex2svg(math_part.math, {display: math_part.display}).outerHTML)
-
-    return math_part.end.n
-  }, 0)
-
-  return parts.join("")
-}
-
-export function contains_tex_string(text: unknown): boolean {
-  if (!isString(text)) return false
-
-  return Boolean(find_math(text))
-};
 
 export function is_tex_string(text: unknown): boolean {
   if (!isString(text)) return false
